fix(embedqueue): match both flock errnos, gate lock tests to unix, correct OnError contract doc
